Backup Themes: Prevent us from getting stuck when the system checks backup names
Scenario: The user enters a name that already exists. Then he changes
his input and chooses a name that does not yet exist. Since the variable
mThemeNameExist is still set to true from the last search, every further
user entry is reported with "Name already exists" and the user can only
cancel the dialog to continue.

Change-Id: I6009d71f6e01db4eaa3ba954c8a2d9bc6f41323f